added FriendForm.js, transferred code from App.js
