+++ friends/src/index.js
@@ -1,13 +1,14 @@
-import React from 'react';
-import ReactDOM from 'react-dom';
-import { BrowserRouter as Router } from 'react-router-dom';
-import './index.css';
-import App from './App';
-import registerServiceWorker from './registerServiceWorker';
+import React from "react";
+import ReactDOM from "react-dom";
+import { BrowserRouter as Router } from "react-router-dom";
+import "./index.css";
+import App from "./App";
+import registerServiceWorker from "./registerServiceWorker";
 
 ReactDOM.render(
-	<Router>
-		<App />
-	</Router>,
-	document.getElementById('root'));
+  <Router>
+    <App />
+  </Router>,
+  document.getElementById("root")
+);
 registerServiceWorker();
